requirements.txt
numpy
open3d
trimesh
spatialmath-python==1.0.5
tqdm
opencv-python
Pillow
pytorch-lightning
wandb
black
zstandard
scikit-image
scikit-learn
tyro
pyrealsense2
rerun-sdk
pycocotools
sapien
mplib
roma
shortuuid
colour
fvcore
git+ssh://git@github.com/chisarie/urdfpy#egg=urdfpy
git+ssh://git@github.com/chisarie/mesh_to_sdf#egg=mesh_to_sdf